backend/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
pydantic
python-jose[cryptography]
passlib[bcrypt]
psutil
